dev_requirements.txt
# This file is used for testing and development. See
# https://caremad.io/posts/2013/07/setup-vs-requirement/

# Install this package
-e .

# Install testing / development requirements
pytest
pytest-django
mock
fakeredis
